add authentication

+++ src/main/java/com/example/JournalApp/Service/JournalEntryService.java
@@ -3,15 +3,21 @@ package com.example.JournalApp.Service;
 import com.example.JournalApp.Entity.Journal;
 import com.example.JournalApp.Entity.User;
 import com.example.JournalApp.Repository.JournalEntryRepository;
+import lombok.extern.slf4j.Slf4j;
+import org.bson.types.ObjectId;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
-@Component
+@Service
+@Slf4j
 public class JournalEntryService {
 
     @Autowired
@@ -27,11 +33,11 @@ public class JournalEntryService {
             journal.setDate(LocalDateTime.now());
             Journal saved =  journalEntryRepository.save(journal);
             user.getJournalEntries().add(saved);
-            user.setUserName(null);
-            userService.saveEntry(user);
+//            user.setUserName(null);
+            userService.saveUser(user);
         } catch (Exception e){
-            System.out.println(e);
-            throw new RuntimeException("An error occur while saving the entry", e);
+//            System.out.println(e);
+            throw new RuntimeException("An error occur while saving the entry.", e);
         }
     }
 
@@ -43,15 +49,25 @@ public class JournalEntryService {
         return journalEntryRepository.findAll();
     }
 
-    public Optional<Journal> fetchJournalById(Long journalId){
+    public Optional<Journal> fetchJournalById(ObjectId journalId){
         return journalEntryRepository.findById(journalId);
     }
 
-    public void deleteById(Long journalId, String userName){
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getJournalId().equals(journalId));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(journalId);
+    @Transactional
+    public boolean deleteById(ObjectId journalId, String userName){
+        boolean removed = false;
+        try{
+            User user = userService.findByUserName(userName);
+            removed = user.getJournalEntries().removeIf(x -> x.getJournalId().equals(journalId));
+            if (removed){
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(journalId);
+            }
+        } catch (Exception e){
+            log.error("Error ",e);
+            throw new RuntimeException("An error occurred while deleting the entry.", e);
+        }
+        return removed;
     }
 
 }
